Simplify app creation
