+++ src/index.ts
@@ -1,3 +1,24 @@
-import app from './app';
+import cors from '@koa/cors';
+import Koa from 'koa';
+import logger from 'koa-logger';
+import InMemoryArticles from './adaptors/in-memory-articles';
+import Articles from './articles';
+import apiDocumentation from './middleware/api-documentation';
+import errorHandler from './middleware/error-handler';
+import routing from './middleware/routing';
+import createRouter from './router';
 
-app.listen(8080);
+const app = new Koa();
+
+const articles: Articles = new InMemoryArticles();
+const router = createRouter(articles);
+
+app.use(logger());
+app.use(cors({
+  exposeHeaders: ['Link'],
+}));
+app.use(apiDocumentation(router));
+app.use(errorHandler());
+app.use(routing(router));
+
+export default app.listen(8080);
